better abstract loading rom files
