+++ src/memory/mod.rs
@@ -3,14 +3,12 @@ mod branch;
 mod null;
 pub mod pia;
 pub mod ports;
-mod romfile;
 pub mod via;
 
 pub use block::BlockMemory;
 pub use branch::BranchMemory;
 pub use null::NullMemory;
 pub use ports::{NullPort, Port};
-pub use romfile::RomFile;
 
 /// Represents the state of the interrupts on the system.
 pub enum ActiveInterrupt {
