+++ src/systems/brooke.rs
@@ -1,6 +1,7 @@
 use crate::memory::{ActiveInterrupt, Memory, SystemInfo};
-use crate::memory::{BlockMemory, BranchMemory, RomFile};
+use crate::memory::{BlockMemory, BranchMemory};
 use crate::platform::PlatformProvider;
+use crate::roms::RomFile;
 use crate::system::System;
 use crate::systems::SystemFactory;
 use std::sync::Arc;
