+++ src/systems/klaus.rs
@@ -1,5 +1,6 @@
-use crate::memory::{BlockMemory, BranchMemory, RomFile};
+use crate::memory::{BlockMemory, BranchMemory};
 use crate::platform::PlatformProvider;
+use crate::roms::RomFile;
 use crate::system::System;
 use crate::systems::SystemFactory;
 use std::sync::Arc;
